feat: stickyTabBar 컴포넌트 분리
