+++ src/pages/templeDetailPage/TempleDetailPage.tsx
@@ -2,6 +2,7 @@ import DetailCarousel from '@components/carousel/detailCarousel/DetailCarousel';
 import ButtonBar from '@components/common/button/buttonBar/ButtonBar';
 import TapBar from '@components/common/tapBar/TapBar';
 import SmallMap from '@components/templeDetail/naverMap/smallMap/SmallMap';
+import StickyTapBar from '@components/templeDetail/stickyTapBar/StickyTapBar';
 import TempleDetailInfo from '@components/templeDetail/templeDetailInfo/TempleDetailInfo';
 import TempleInfo from '@components/templeDetail/templeInfo/templeInfo';
 import TemplePrice from '@components/templeDetail/templePrice/TemplePrice';
@@ -9,36 +10,10 @@ import TempleReview from '@components/templeDetail/templeReview/TempleReview';
 import TempleSchedule from '@components/templeDetail/templeSchedule/TempleSchedule';
 import TempleTitle from '@components/templeDetail/templeTitle/TempleTitle';
 import TempleTopbar from '@components/templeDetail/templeTopbar/TempleTopbar';
-import { useEffect, useRef, useState } from 'react';
 
 import * as styles from './templeDetailPage.css';
 
 const TempleDetailPage = () => {
-  const tapBarRef = useRef<HTMLDivElement>(null);
-  const [isSticky, setIsSticky] = useState(false);
-  const [initialOffsetTop, setInitialOffsetTop] = useState<number | null>(null);
-
-  useEffect(() => {
-    if (tapBarRef.current) {
-      setInitialOffsetTop(tapBarRef.current.getBoundingClientRect().top);
-    }
-  }, []);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      if (initialOffsetTop === null) return;
-
-      const scrollTop = window.scrollY;
-
-      setIsSticky(scrollTop >= initialOffsetTop - 62);
-    };
-
-    window.addEventListener('scroll', handleScroll);
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, [initialOffsetTop]);
-
   return (
     <div className={styles.templeDetailWrapper}>
       <div className={styles.headerBox}>
@@ -50,9 +25,9 @@ const TempleDetailPage = () => {
           <TempleTitle />
           <TempleDetailInfo />
         </div>
-        <div ref={tapBarRef} className={`${isSticky ? styles.stickyTapBar : ''}`}>
+        <StickyTapBar>
           <TapBar type="detail" />
-        </div>
+        </StickyTapBar>
       </div>
       <div className={styles.templeDetailMiddle}>
         <TempleReview />
